# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: NPR Seminar NestJS API
  version: 1.0.0
  description: |
    Authoritative contract for the single NestJS backend used by the NPR Seminar
    web and scanner clients.

    ADMIN and SCANNER actors use a Redis-backed project session cookie. Every
    state-changing session request requires X-CSRF-Token, an allowed same-origin
    Origin, and application/json. Public family access uses a separate,
    contact-bound X-Booking-Proof credential. QR tokens, booking proofs, OTPs,
    cookies, passwords, pairing codes, and full parent contact values are secrets and
    MUST be redacted from application, proxy, tracing, analytics, and audit logs.

    Durable resources are never hard-deleted. Booking, check-in, student, sync,
    circuit, and scanner-device histories are append-only and returned in stable
    chronological order.
  x-contract-authority: packages/contracts/openapi.yaml
  x-backend-runtime: NestJS modular monolith
  x-sensitive-log-redaction:
    - Cookie
    - Set-Cookie
    - X-CSRF-Token
    - X-Booking-Proof
    - bookingProof
    - X-QR-Token
    - qrToken
    - pairingCode
    - oneTimeCode
    - contact
    - motherPhone
    - fatherPhone
    - password
  x-canonical-student-snapshot:
    snapshotDate: 2026-07-17
    classNameAdmissionPolicy:
      normalizeForDecision: NFKC_TRIM
      preserveAcceptedSourceValue: true
      allowBracketless: true
      allowOneTerminalScheduleSuffix: true
...
      rejectAsterisk: true
...
    scope: ALL
...
servers:
...
paths:
...
  /api/v1/public/students:
    get:
      operationId: searchAuthorizedStudents
      tags: [Public students]
      summary: Search students authorized by the contact-bound booking proof
      description: |
        The server derives the parent-contact digest only from X-Booking-Proof
        and returns active students in the campus bound into that proof whose
        mother or father contact digest exactly matches it. The optional branch
        query is compatibility-only and, when present, must equal the proof-bound
        campus. A caller cannot broaden this set with query parameters or a
        guessed student ID.
      security:
        - bookingProof: []
      x-proof-scopes: [STUDENT_SEARCH, FAMILY_BOOKING]
      parameters:
        - name: query
          in: query
          required: false
          description: Optional name or source student number prefix inside the proof-authorized set.
          schema:
            type: string
            minLength: 1
            maxLength: 80
        - $ref: '#/components/parameters/BranchQuery'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Proof-authorized students only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicStudentPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v1/admin/seminar-sessions/{seminarSessionId}/survey-responses:
    parameters:
      - $ref: '#/components/parameters/SeminarSessionId'
    get:
      operationId: listSessionSurveyResponses
      tags: [Admin seminars]
      summary: List safe survey responses and rating summary for one session
      description: |
        Responses are ordered newest first. Each ADMIN-only item includes a
        representative current participant context and the full booking contact
        needed for operational follow-up. Active booking links win; if none are
        active, the latest released participant set is used. An active current
        enrolled student wins over another enrolled snapshot, which wins over a
        guest. participantCount and additionalParticipantCount retain the family
        context when one representative sibling is displayed.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Session survey response page and aggregate summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSurveyResponsePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/students:
    get:
      operationId: listAdminStudents
      tags: [Admin students]
      summary: List students with source assignments and representative resolution
      description: |
        Source assignments are preserved in full. Representative candidates use
        the corrected rules after stripping one validated terminal timetable
        suffix: science when the base className starts with 과 or names a
        science subject such as 물리, 화학, 생명과학/생물, or 지구과학; science
        displays as 과학. Names containing 특강, 패키지, 입시대비, or TEST
        case-insensitively are supplementary and excluded only from representative
        candidates. Accepted suffixes are distinct weekday sequences of one to
        three characters with an optional positive 1..99 slot without a leading
        zero, or E/e with that slot. Asterisks and every other bracket form are
        rejected. 기하[시간] is retained as subject-only, not representative.
        고3이과T is regular. One regular wins; zero regular plus one
        or more science becomes 과학; multiple regular or no regular/science is
        ambiguous. Active representative assignments are projected once for
        display, grouping, and ordering. Math is every eligible non-science
        assignment; science-only means eligible science and no math. Students
        with both retain both projections, and the unfiltered list excludes
        neither category. Projected class names preserve the exact NFKC-normalized
        raw timetable suffix and de-duplicate exact names only. Results are sorted
        before pagination by branch CAMPUS_A, CAMPUS_B, CAMPUS_C; then projected unit,
        projected class, student name, and source student number ascending.
        facets.teachers contains distinct primary teacher names (NFKC-normalized,
        comma-delimited source values reduced to the first name) across the full
        non-paginated result for every current filter except teacherName, but only
        for students whose active classification projection has at least one math
        class. Science-only and otherwise no-math students never contribute a
        teacher facet and can never match teacherName. The selected teacher
        continues to filter items and page totals but never narrows its own option
        list. summary is intentionally separate: all
        fields except reviewRequiredStudentCount aggregate sourceActive=true
        students for only the selected branch and unitGroup and ignore every
        other list filter. reviewRequiredStudentCount uses sourceActive=true and
        branch only, ignoring unitGroup and every other list filter.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      parameters:
        - $ref: '#/components/parameters/BranchQuery'
        - name: query
          in: query
          required: false
          description: Allowlisted search over name, school, source student number, or approved contact lookup.
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: grade
          in: query
          required: false
          deprecated: true
          description: Compatibility-only legacy filter. New clients use unitGroup; this parameter will be removed after the transition window.
          schema:
            type: string
            maxLength: 40
        - name: unitGroup
          in: query
          required: false
          description: Same canonical unit grouping as the reservation roster. GUEST returns an empty student-status page because guests are not synchronized students.
          schema:
            $ref: '#/components/schemas/RosterUnitGroup'
        - name: representativeClass
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
        - name: teacherName
          in: query
          required: false
          description: Exact math-homeroom teacher filter. NFKC normalization is applied and only the first comma-delimited student-row name is used. Science-only and other no-math students never match.
          schema:
            type: string
            maxLength: 100
        - name: unitName
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
        - name: resolution
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/RepresentativeResolution'
        - name: sourceActive
          in: query
          required: false
          schema:
            type: boolean
        - name: categorizedOnly
          in: query
          required: false
          description: When true, items, page totals, and teacher facets include only students with at least one recognizable active math or science class. False or omission does not apply this filter. Classification summary fields ignore it.
          schema:
            type: boolean
        - name: seminarSessionId
          in: query
          required: false
          description: When supplied, adds one booking projection for each page student in this seminar session, prioritizing an active RESERVED or CHECKED_IN booking and otherwise selecting the latest historical booking. The projection is fetched in one page-level batch. When omitted or no booking matches, reservation is null and hasReservation is false. hasReservation is true for RESERVED, CHECKED_IN, and NO_SHOW, and false for CANCELLED.
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Student page, page-independent primary-teacher facets, and independently reported multi-assignment/classification summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStudentPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/students/review-required:
    get:
      operationId: listAdminStudentsReviewRequired
      tags: [Admin students]
      summary: List active students requiring assignment review
      description: |
        Returns sourceActive=true students with one or more deterministic anomaly
        reasons: MULTIPLE_MATH_CLASS, NO_RECOGNIZABLE_CLASS, UNIT_UNRESOLVED,
        or ABNORMAL_OR_EMPTY_CLASS. Campus is exposed as branch. Each item has
        exactly studentId, sourceStudentNo, branch, name, originalClassName,
        mathClassName, scienceClassNames, reasonCodes, rawClassNames, and
        rawRepresentativeClassNames. rawClassNames preserves every distinct
        current active assignment class value verbatim, including invalid,
        whitespace-only, and empty values, so source context is not lost.
        rawRepresentativeClassNames remains the rolling-compatibility projection
        of exact NFKC-normalized recognizable names. The response is exactly
        {items,page}; page.totalItems is the authoritative filtered total and
        must not be reconstructed from items.length.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      x-authoritative-total-field: page.totalItems
      parameters:
        - $ref: '#/components/parameters/BranchQuery'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paginated students requiring review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStudentReviewRequiredPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/students/{studentId}:
    parameters:
      - $ref: '#/components/parameters/StudentId'
    get:
      operationId: getAdminStudent
      tags: [Admin students]
      summary: Get a student with every current source assignment
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      responses:
        '200':
          description: Student detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStudentDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/students/{studentId}/history:
    parameters:
      - $ref: '#/components/parameters/StudentId'
    get:
      operationId: listAdminStudentHistory
      tags: [Admin students]
      summary: List append-only student and assignment audit history
      description: Events are immutable and ordered by sequence ascending, then occurredAt ascending.
      security:
        - cookieAuth: []
      x-required-roles: [ADMIN]
      parameters:
        - $ref: '#/components/parameters/AfterSequence'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Chronological student history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentAuditEventPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: npr_seminar_session
      description: |
        Opaque project-specific Redis session cookie for ADMIN and SCANNER.
        Production attributes are HttpOnly, Secure, SameSite=Lax, Path=/, with
        no Domain where possible. Cookie values are never logged.
    bookingProof:
      type: apiKey
      in: header
      name: X-Booking-Proof
      description: |
        Short-lived 256-bit base64url bearer credential bound to a normalized
        family contact digest, purpose-derived scopes, and expiry. PostgreSQL
        stores SHA-256(proof) only. The raw value is returned once, is independent
        from the project session cookie, and must be redacted from every log.
    qrCredential:
      type: apiKey
      in: header
      name: X-QR-Token
      description: |
        Opaque family QR credential used only for the minimal public pass.
        The raw value is secret, never persisted, and redacted from all logs.
...
  parameters:
...
    SeminarSessionId:
      name: seminarSessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    StudentId:
      name: studentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    BranchQuery:
      name: branch
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/Branch'
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    AfterSequence:
      name: afterSequence
      in: query
      required: false
      description: Return immutable events with sequence greater than this value.
      schema:
        type: string
        pattern: '^[0-9]+$'
        default: '0'
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    BadRequest:
      description: Malformed input, invalid header, or invalid confirmation
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing, expired, invalid, or revoked session/proof/credential
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Role, CSRF, same-origin, contact ownership, or scope check failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found or intentionally concealed from this actor
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    RateLimited:
      description: Request rate limit exceeded
      headers:
        Retry-After:
          description: Seconds until this request class may be retried
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ServiceUnavailable:
      description: Required PostgreSQL, Redis, source, or cryptographic configuration is unavailable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details. Secret or upstream raw values are forbidden in every field.
      required: [type, title, status, detail, instance, code, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{2,80}$'
        traceId:
          type: string
          minLength: 16
          maxLength: 128
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ProblemFieldError'
      additionalProperties: true
...
    Branch:
      type: string
      enum: [CAMPUS_A, CAMPUS_B, CAMPUS_C]
      x-labels:
        CAMPUS_A: A
        CAMPUS_B: B
        CAMPUS_C: C
...
    PageMeta:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
      additionalProperties: false
...
    RepresentativeResolution:
      type: string
      enum: [REGULAR, SCIENCE_ALIAS, MULTIPLE_REGULAR, NO_CLASS]
      description: |
        REGULAR selects exactly one regular candidate. SCIENCE_ALIAS applies when
        there are zero regular and at least one science candidate. The remaining
        two values are ambiguous.
...
    PublicStudent:
      type: object
      required:
        - studentId
        - sourceStudentNo
        - name
        - branch
        - schoolName
        - grade
        - representativeClass
      properties:
        studentId:
          type: string
          format: uuid
        sourceStudentNo:
          type: string
        name:
          type: string
        branch:
          $ref: '#/components/schemas/Branch'
        schoolName:
          type: [string, 'null']
        grade:
          type: [string, 'null']
        representativeClass:
          $ref: '#/components/schemas/RepresentativeClass'
      additionalProperties: false
    PublicStudentPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PublicStudent'
        page:
          $ref: '#/components/schemas/PageMeta'
      additionalProperties: false
    AdminStudent:
      type: object
      required:
        - studentId
        - sourceStudentNo
        - name
        - branch
        - schoolName
        - grade
        - teacherName
        - motherPhone
        - fatherPhone
        - unitName
        - sourceActive
        - mathClassName
        - scienceClassNames
        - hasReservation
        - reservation
        - representativeClass
        - assignments
        - firstSeenAt
        - lastSeenAt
      properties:
        studentId:
          type: string
          format: uuid
        sourceStudentNo:
          type: string
          minLength: 1
          maxLength: 100
        name:
          type: string
          minLength: 1
          maxLength: 100
        branch:
          $ref: '#/components/schemas/Branch'
        schoolName:
          type: [string, 'null']
          maxLength: 200
        grade:
          type: [string, 'null']
          maxLength: 40
        teacherName:
          type: [string, 'null']
          maxLength: 100
          description: |
            Required and nullable math-homeroom teacher. When mathClassName is
            non-null, this is the NFKC-normalized first comma-separated value
            from students.teacher_name. When mathClassName is null, including
            for science-only students, this is null even when the student row
            or assignments contain teacher names. StudentSourceAssignment
            teacherName values are traceability only and never homeroom sources
            or fallbacks.
          x-required-nullable: true
          x-homeroom-applicability: MATH_REGULAR_CLASS_ONLY
          x-source-field: students.teacher_name
          x-source-condition: MATH_CLASS_PRESENT
          x-normalization: NFKC_TRIM_FIRST_COMMA_VALUE
          x-science-only-policy: RETURN_NULL
          x-no-math-policy: RETURN_NULL
          x-assignment-teacher-role: TRACEABILITY_ONLY
          x-assignment-teacher-fallback: forbidden
        motherPhone:
          type: [string, 'null']
          minLength: 8
          maxLength: 15
          pattern: '^[0-9]{8,15}$'
          description: Full normalized mother phone; ADMIN response only and forbidden from logs and audit metadata.
        fatherPhone:
          type: [string, 'null']
          minLength: 8
          maxLength: 15
          pattern: '^[0-9]{8,15}$'
          description: Full normalized father phone; ADMIN response only and forbidden from logs and audit metadata.
        unitName:
          type: [string, 'null']
          enum: [과학, 예고1, 특목, 예중1, 초등, 고등, 중등1, 중등2, 중등3, null]
          maxLength: 100
        sourceActive:
          type: boolean
        mathClassName:
          type: [string, 'null']
          maxLength: 200
          description: Preferred exact NFKC-normalized active representative non-science class, including a valid raw timetable suffix. The persisted representative class wins when its exact normalized name is eligible; otherwise the stable first eligible name is used.
        scienceClassNames:
          type: array
          description: Exact-name-de-duplicated stable order of every NFKC-normalized active representative science class. Valid raw timetable suffixes are preserved.
          items:
            type: string
            minLength: 1
            maxLength: 200
        hasReservation:
          type: boolean
          description: True when the selected seminar-session projection has status RESERVED, CHECKED_IN, or NO_SHOW. False for CANCELLED, when no matching booking exists, and when seminarSessionId is omitted.
        reservation:
          description: Selected seminar-session booking projection, or null when seminarSessionId is omitted or the student has no booking in that session.
          anyOf:
            - $ref: '#/components/schemas/AdminStudentReservation'
            - type: 'null'
        representativeClass:
          $ref: '#/components/schemas/RepresentativeClass'
        assignments:
          type: array
          description: All current source assignments, including supplementary assignments.
          items:
            $ref: '#/components/schemas/StudentSourceAssignment'
        firstSeenAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
      additionalProperties: false
    AdminStudentDetail:
      $ref: '#/components/schemas/AdminStudent'
...
    AdminStudentReviewRequiredPage:
      type: object
      description: Exactly items and page; page.totalItems is the authoritative filtered review count.
      x-authoritative-total-field: page.totalItems
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminStudentReviewRequired'
        page:
          $ref: '#/components/schemas/PageMeta'
      additionalProperties: false
...
    AdminStudentFacets:
      type: object
      description: |
        Filter options computed before pagination. teachers applies every current
        admin-student filter except teacherName, page, and pageSize, so selecting
        a teacher filters the result without collapsing the available options.
        Values are distinct NFKC-normalized primary names taken before the first
        comma and returned in stable ascending order, only for students with at
        least one active projected math class. Science assignment instructors are
        traceability-only and science-only students are excluded.
      required: [teachers]
      properties:
        teachers:
          type: array
          uniqueItems: true
          x-homeroom-applicability: MATH_REGULAR_CLASS_ONLY
          x-science-only-policy: EXCLUDED
          items:
            type: string
            minLength: 1
            maxLength: 100
      additionalProperties: false
    AdminStudentPage:
      type: object
      required: [items, page, summary, facets, latestSuccessfulSyncAt]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminStudent'
        page:
          $ref: '#/components/schemas/PageMeta'
        summary:
          $ref: '#/components/schemas/StudentClassificationSummary'
        facets:
          $ref: '#/components/schemas/AdminStudentFacets'
        latestSuccessfulSyncAt:
          type: [string, 'null']
          format: date-time
      additionalProperties: false
...
    RosterUnitGroup:
      type: string
      enum: [ALL, ELEMENTARY, MIDDLE_1, MIDDLE_2, MIDDLE_3, SPECIAL_PURPOSE, HIGH, SCIENCE, GUEST]
...
    SurveyResponse:
      type: object
      required:
        - surveyResponseId
        - familyBookingId
        - participant
        - rating
        - comment
        - photoAttached
        - photoName
        - submittedAt
      properties:
        surveyResponseId:
          type: string
          format: uuid
        familyBookingId:
          type: string
          format: uuid
        participant:
          $ref: '#/components/schemas/SurveyParticipantContext'
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: [string, 'null']
          maxLength: 2000
        photoAttached:
          type: boolean
        photoName:
          type: [string, 'null']
          maxLength: 255
        submittedAt:
          type: string
          format: date-time
      additionalProperties: false
...
    SurveyResponseSummary:
      type: object
      required: [averageRating, responseCount, ratingDistribution]
      properties:
        averageRating:
          type: [number, 'null']
          minimum: 1
          maximum: 5
        responseCount:
          type: integer
          minimum: 0
        ratingDistribution:
          $ref: '#/components/schemas/SurveyRatingDistribution'
      additionalProperties: false
    AdminSurveyResponsePage:
      type: object
      required: [items, page, summary]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SurveyResponse'
        page:
          $ref: '#/components/schemas/PageMeta'
        summary:
          $ref: '#/components/schemas/SurveyResponseSummary'
      additionalProperties: false
...
    StudentAuditEvent:
      type: object
      description: Append-only student identity, assignment, and resolution history.
      required:
        - eventId
        - sequence
        - studentId
        - runId
        - type
        - assignment
        - representativeClass
        - changes
        - occurredAt
      properties:
        eventId:
          type: string
          format: uuid
        sequence:
          $ref: '#/components/schemas/AuditSequence'
        studentId:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/StudentAuditEventType'
        assignment:
          oneOf:
            - $ref: '#/components/schemas/StudentSourceAssignment'
            - type: 'null'
        representativeClass:
          oneOf:
            - $ref: '#/components/schemas/RepresentativeClass'
            - type: 'null'
        changes:
          type: array
          items:
            $ref: '#/components/schemas/StudentFieldChange'
        occurredAt:
          type: string
          format: date-time
      additionalProperties: false
    StudentAuditEventPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          description: Strict sequence-ascending chronology.
          items:
            $ref: '#/components/schemas/StudentAuditEvent'
        page:
          $ref: '#/components/schemas/EventPageMeta'
      additionalProperties: false
...